% Auto-generated: do not edit by hand
\name{dashSocketIO}

\alias{dashSocketIO}

\title{DashSocketIO component}

\description{

}

\usage{
dashSocketIO(id=NULL, debug=NULL, send=NULL, url=NULL)
}

\arguments{
\item{id}{Character. Unique ID to identify this component in Dash callbacks.}

\item{debug}{Logical. }

\item{send}{Named list. }

\item{url}{Character. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

